consolidate UUID generation into shared generateId util with Math.random fallback
