+++ frontend/editor/src/core/utils/browserIdentifier.ts
@@ -3,6 +3,8 @@
  * Generates and persists a unique UUID in localStorage for WAU tracking
  */
 
+import { generateId } from "@app/utils/generateId";
+
 const BROWSER_ID_KEY = "stirling_browser_id";
 
 /**
@@ -28,19 +30,6 @@ export function getBrowserId(): string {
   }
 }
 
-/**
- * Generates a UUID v4
- */
 function generateUUID(): string {
-  // Use crypto.randomUUID if available (modern browsers)
-  if (typeof crypto !== "undefined" && crypto.randomUUID) {
-    return crypto.randomUUID();
-  }
-
-  // Fallback to manual UUID generation
-  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
-    const r = (Math.random() * 16) | 0;
-    const v = c === "x" ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
+  return generateId();
 }
